HACK: domd: weston: Add suport multi domains and display
weston doesn't support window control so that it is impossible to place
application on specifc display automatically.
This is inconvenient for demonstration with multi display

This patch modify the kiosk-shell to run application to each display.
By this modification, Application runs on next display by every
launching.
I think this is enough for demonstration purpose.

Signed-off-by: Yuya Hamamachi <yuya.hamamachi.sx@renesas.com>